New Error Pages + OAMPlus Requirement
Error Pages are converted to .js so we dont have to redirect. Saves some time and makes it friendly to reload if Socket server is ever to have temporary downtime.

Plus those who dont claim a client ID will be stuck on install proccess until done. This is to combat a bug back in 2.x spamming the alert.

Signed-off-by: Legoman99573 <christianwhead@gmail.com>